% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.report.june.2022.helper.R
\name{get_pm2.5_le_gains}
\alias{get_pm2.5_le_gains}
\title{Get the average PM2.5 pollution and corresponding LE gains of a given region}
\usage{
get_pm2.5_le_gains(
  df = NULL,
  gadm_0 = NULL,
  gadm_1 = NULL,
  gadm_2 = NULL,
  year = 2020,
  red_to = 5,
  le_constant = 0.098
)
}
\arguments{
\item{df}{master color dataset}

\item{gadm_0}{country name, which corresponds to gadm level 0 column in the gadm shape file.
This can also be a vector of length > 1, but in that case both gadm_1 and gadm_2 vector length should be <= 1.}

\item{gadm_1}{corresponds to the gadm level 1 region in the gadm shape file.This can also be a vector of length > 1,
but in that case both gadm_1 and gadm_2 vector length should be <= 1.}

\item{gadm_2}{corresponds to gadm level 2 region in the gadm shape file. This can also be a vector of length > 1,
but in that case both gadm_1 and gadm_2 vector length should be <= 1.}

\item{year}{the year for which we want the average PM2.5 and life expectancy gains for the region in question, deafults to 2020}

\item{red_to}{the amount to which pm2.5 pollution should be reduced, by deafult this is the 5 micrograms per cubic meter}

\item{le_constant}{its the constant with which we multiply the pm2.5 concentrations value by to get the life expectancy gains in years}
}
\value{
a summary table with an average PM2.5 column and the corresponding Life Expectancy Gains column
}
\description{
This function gets the average pollution and LE gains number for a given gadm level 0, 1 and 2 region in a given year
}
\examples{

get_pm2.5_le_gains(color_2020, gadm_0 == "India") # gives India's average PM2.5 and LE gains for India for 2020


}
