% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annual.report.june.2022.helper.R
\name{get_pm2.5_le_gains_all_sub_regions}
\alias{get_pm2.5_le_gains_all_sub_regions}
\title{Get all sub-regions average pm2.5 pollution and life expectancy gains within a specified region}
\usage{
get_pm2.5_le_gains_all_sub_regions(
  df,
  global = FALSE,
  gadm_0 = NULL,
  gadm_1 = NULL,
  gadm_2 = NULL,
  level = NULL,
  year = 2020,
  red_to = 5,
  le_constant = 0.098
)
}
\arguments{
\item{df}{master color dataset}

\item{global}{defaults to FALSE, but if TRUE, it gives us a average for all countries (which are global sub-regions)}

\item{gadm_0}{defaults to NULL, but if not NULL it gives us the average PM2.5 and le gains of the gadm level 1 regions of the gadm_0 region specified}

\item{gadm_1}{defaults to NULL, but if not NULL it gives us the average PM2.5 and le gains of the gadm level 2 regions of the gadm_0 and gadm_1 regions specified.}

\item{gadm_2}{defaults to NULL, but if not NULL it gives us the average PM2.5 and le gains of the gadm level 2 region, which is present in the gadm_1 region,
which itself is present in the  gadm_0 region.}

\item{level}{This can be either "gadm_1" or "gadm_2", and applies only to the case when gadm_0 information is needed.
This determines the level at which the output will be supplied}

\item{year}{this is the year in question for which we need the average PM2.5 and le gains.}

\item{red_to}{what's the threshold relative to which we want to calculate the le gains, default to WHO guideline of 5 micrograms per cubic meter}

\item{le_constant}{defaults to 0.098, this stays the same. It is the finding from the recent AQLI paper.}
}
\value{
summary dataset containing average PM2.5 and le gains for the region in question.
}
\description{
This function calculates the average PM2.5 and corresponding life expectancy gains for all sub-regions within
a given specified region.
}
\note{
Please note that this function provides the average PM2.5 and LE gains for all sub-regions
      within the region specified in the function's parameters.
}
